+++ src/BaseServiceProvider.php
@@ -18,22 +18,34 @@ class BaseServiceProvider extends ServiceProvider
         // LOAD THE VIEWS
         // - first the published views (in case they have any changes)
         $this->loadViewsFrom(resource_path('views/vendor/bytenet/laravel-admin-base'), 'bytenet');
+        // - then the stock views that come with the package, in case a published view might be missing
+        $this->loadViewsFrom(realpath(__DIR__ . '/resources/views'), 'bytenet');
 
-        $this->loadTranslationsFrom(realpath(__DIR__.'/resources/lang'), 'bytenet');
+        // use the vendor translation file as fallback
+        $this->loadTranslationsFrom(realpath(__DIR__ . '/resources/lang'), 'bytenet');
+        
+        // use the vendor configuration file as fallback
+        $this->mergeConfigFrom(__DIR__ . '/config/bytenet/base.php', 'bytenet.base');
 
-
-        $this->setupRoutes($this->app->router);
+        $this->map();
 
         // -------------
         // PUBLISH FILES
         // -------------
         // publish config file
         $this->publishes([__DIR__.'/config' => config_path()], 'config');
+        // publish migrations
+        $this->publishes([__DIR__.'/database/migrations' => database_path('migrations')], 'migrations');
+        // publish seed
+        $this->publishes([__DIR__.'/database/seeds' => database_path('seeds')], 'seeds');
         // publish lang files
         $this->publishes([__DIR__.'/resources/lang' => resource_path('lang/vendor/bytenet')], 'lang');
 
         // publish views
-        $this->publishes([__DIR__.'/resources/views' => resource_path('views/vendor/bytenet/laravel-admin-base')], 'views');
+        $this->publishes(
+            [__DIR__.'/resources/views' => resource_path('views/vendor/bytenet/laravel-admin-base')],
+            'views'
+        );
     }
 
     /**
@@ -43,7 +55,10 @@ class BaseServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        //
+        // register the current package
+        $this->app->bind('bytenet.admin.base', function ($app) {
+            return new Base($app);
+        });
     }
 
 
@@ -54,28 +69,45 @@ class BaseServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function setupRoutes(Router $router)
+    public function map()
     {
         // register the 'admin' middleware
-        $router->middleware('auth.bytenet', app\Http\Middleware\AuthenticateByteNet::class);
+        Route::middleware('bytenet.auth', app\Http\Middleware\ByteNetAuthenticate::class);
 
-        $router->group(['namespace' => 'ByteNet\LaravelAdminBase\app\Http\Controllers'], function ($router) {
-            Route::group(['middleware' => 'web', 'prefix' => config('bytenet.base.route_prefix')], function () {
+        $this->mapWebRoutes();
 
-                // if not otherwise configured, setup the auth routes
-                if (config('bytenet.base.setup_auth_routes')) {
-                    Route::auth();
-                }
+        $this->mapApiRoutes();
+    }
 
-                // if not otherwise configured, setup the dashboard routes
-                if (config('bytenet.base.setup_dashboard_routes')) {
-                    Route::get('dashboard', 'AdminController@dashboard');
-                    Route::get('/', function () {
-                        // The '/admin' route is not to be used as a page, because it breaks the menu's active state.
-                        return redirect(config('bytenet.base.route_prefix').'/dashboard');
-                    });
-                }
-            });
+
+    /**
+     * Define the "web" routes for the application.
+     *
+     * These routes all receive session state, CSRF protection, etc.
+     *
+     * @return void
+     */
+    protected function mapWebRoutes()
+    {
+        Route::group([
+            'middleware' => 'web',
+            'prefix' => config('bytenet.base.route_prefix'),
+            'namespace' => 'ByteNet\LaravelAdminBase\app\Http\Controllers',
+            'as' => config('bytenet.base.route_prefix').'::',
+        ], function ($router) {
+            require __DIR__ . '/routes/web.php';
         });
+    }
+
+    /**
+     * Define the "api" routes for the application.
+     *
+     * These routes are typically stateless.
+     *
+     * @return void
+     */
+    protected function mapApiRoutes()
+    {
+        //
     }
 }