+++ src/resources/views/layout.blade.php
@@ -9,26 +9,53 @@
     <meta name="csrf-token" content="{{ csrf_token() }}" />
 
     <title>
-      {{ isset($title) ? $title.' :: '.config('bytenet.base.project_name').' Admin' : config('bytenet.base.project_name').' Admin' }}
+        {{ isset($title) ? $title.' :: '.config('bytenet.base.project_name').' Admin' : config('bytenet.base.project_name').' Admin' }}
     </title>
 
+
+    <!-- Tell the browser to be responsive to screen width -->
+    <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
+
     @yield('before_styles')
-    
+
+
+
+
+
 
     {{-- ***** BootstrapCDN.com - START ***** --}}
     <!-- Fonts -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css" integrity="sha384-T8Gy5hrqNKT+hzMclPo118YTQO6cYprQmhrYwIiQ/3axmI1hQomh7Ud2hPOy8SP1" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:100,300,400,700">
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css" integrity="sha384-T8Gy5hrqNKT+hzMclPo118YTQO6cYprQmhrYwIiQ/3axmI1hQomh7Ud2hPOy8SP1" crossorigin="anonymous">
+        <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:100,300,400,700">
 
-    <!-- Latest compiled and minified CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+        <!-- Latest compiled and minified CSS -->
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
 
-    <!-- Optional theme -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
+        <!-- Optional theme -->
+    {{-- <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous"> --}}
 
     <!-- Latest compiled and minified JavaScript -->
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
     {{-- ****** BootstrapCDN.com - END ****** --}}
+
+
+    <!-- Ionicons -->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/ionicons/2.0.1/css/ionicons.min.css">
+    <!-- Theme style -->
+    <link rel="stylesheet" href="/backend/adminlte/dist/css/AdminLTE.min.css">
+    <!-- AdminLTE Skins. We have chosen the skin-blue for this starter
+          page. However, you can choose any other skin. Make sure you
+          apply the skin class to the body tag so the changes take effect.
+    -->
+    <link rel="stylesheet" href="/backend/adminlte/dist/css/skins/skin-blue.min.css">
+
+    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
+    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
+    <!--[if lt IE 9]>
+    <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
+    <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
+    <![endif]-->
+
+
 
     {{-- <link href="{{ elixir('css/app.css') }}" rel="stylesheet" type="text/css"> --}}
 
@@ -39,39 +66,52 @@
 
     
 
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-    <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
-    <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
 </head>
+
+<!--
+BODY TAG OPTIONS:
+=================
+Apply one or more of the following classes to get the
+desired effect
+|---------------------------------------------------------|
+| SKINS         | skin-blue                               |
+|               | skin-black                              |
+|               | skin-purple                             |
+|               | skin-yellow                             |
+|               | skin-red                                |
+|               | skin-green                              |
+|---------------------------------------------------------|
+|LAYOUT OPTIONS | fixed                                   |
+|               | layout-boxed                            |
+|               | layout-top-nav                          |
+|               | sidebar-collapse                        |
+|               | sidebar-mini                            |
+|---------------------------------------------------------|
+-->
+
 <body class="hold-transition {{ config('bytenet.base.skin') }} sidebar-mini">
     <!-- Site wrapper -->
     <div class="wrapper">
 
-      <header class="main-header">
-        <!-- Logo -->
-        <a href="{{ url('') }}" class="logo">
-          <!-- mini logo for sidebar mini 50x50 pixels -->
-          <span class="logo-mini">{!! config('bytenet.base.logo_mini') !!}</span>
-          <!-- logo for regular state and mobile devices -->
-          <span class="logo-lg">{!! config('bytenet.base.logo_lg') !!}</span>
-        </a>
-        <!-- Header Navbar: style can be found in header.less -->
-        <nav class="navbar navbar-static-top" role="navigation">
-          <!-- Sidebar toggle button-->
-          <a href="#" class="sidebar-toggle" data-toggle="offcanvas" role="button">
-            <span class="sr-only">{{ trans('bytenet::base.toggle_navigation') }}</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-          </a>
+        <!-- Main Header -->
+        <header class="main-header">
+            <!-- Logo -->
+            <a href="{{ url('') }}" class="logo">
+              <!-- mini logo for sidebar mini 50x50 pixels -->
+              <span class="logo-mini">{!! config('bytenet.base.logo_mini') !!}</span>
+              <!-- logo for regular state and mobile devices -->
+              <span class="logo-lg">{!! config('bytenet.base.logo_lg') !!}</span>
+            </a>
+            <!-- Header Navbar: style can be found in header.less -->
+            <nav class="navbar navbar-static-top" role="navigation">
+              <!-- Sidebar toggle button-->
+              <a href="#" class="sidebar-toggle" data-toggle="offcanvas" role="button">
+                <span class="sr-only">{{ trans('bytenet::base.toggle_navigation') }}</span>
+              </a>
 
-          @include('bytenet::inc.menu')
-        </nav>
-      </header>
+              @include('bytenet::inc.menu')
+            </nav>
+        </header>
 
       <!-- =============================================== -->
 
@@ -82,58 +122,54 @@
       <!-- Content Wrapper. Contains page content -->
       <div class="content-wrapper">
         <!-- Content Header (Page header) -->
+        <section class="content-header">
          @yield('header')
+        </section>
 
         <!-- Main content -->
         <section class="content">
-
           @yield('content')
-
         </section>
         <!-- /.content -->
       </div>
       <!-- /.content-wrapper -->
 
       <footer class="main-footer">
-        @if (config('bytenet.base.show_powered_by'))
-            <div class="pull-right hidden-xs">
-               {{ trans('bytenet::base.powered_by') }} <a target="_blank" href="https://github.com/ByteNet-Serbia/LaravelAdminBase">Laravel ByteNet</a>
-            </div>
-        @endif
-        {{ trans('bytenet::base.handcrafted_by') }} <a target="_blank" href="{{ config('bytenet.base.developer_link') }}">{{ config('bytenet.base.developer_name') }}</a>.
+          @if (config('bytenet.base.show_powered_by'))
+              <div class="pull-right hidden-xs">
+                  {{ trans('bytenet::base.powered_by') }} <a target="_blank" href="https://github.com/ByteNet-Serbia/LaravelAdminBase">Lara BTNT</a>
+              </div>
+          @endif
+
+            <strong>Copyright &copy; {{ date('Y.') }}</strong>
+              {{ trans('bytenet::base.handcrafted_by') }} <a target="_blank" href="{{ config('bytenet.base.developer_link') }}">{{ config('bytenet.base.developer_name') }}</a>.
+              All rights reserved.
       </footer>
+
+
+        @include('bytenet::inc.sidebar_control')
+
     </div>
     <!-- ./wrapper -->
 
 
     @yield('before_scripts')
 
-    <!-- JavaScripts -->
-    <script src="https://code.jquery.com/jquery-3.1.0.min.js" integrity="sha256-cCueBR6CsyA4/9szpPfrX3s49M9vUU5BgtiJj06wt/s=" crossorigin="anonymous"></script>
-    {{-- <script src="{{ elixir('js/app.js') }}"></script> --}}
-    
+    <!-- REQUIRED JS SCRIPTS -->
 
-    <!-- page script -->
-    <script type="text/javascript">
-        // To make Pace works on Ajax calls
-        $(document).ajaxStart(function() { Pace.restart(); });
+    <!-- jQuery 2.2.3 -->
+    <script src="/backend/adminlte/plugins/jQuery/jquery-2.2.3.min.js"></script>
+    <!-- Bootstrap 3.3.6 -->
+    {{-- <script src="/backend/adminlte/bootstrap/js/bootstrap.min.js"></script> --}}
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
 
-        // Ajax calls should always have the CSRF token attached to them, otherwise they won't work
-        $.ajaxSetup({
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                }
-            });
+    <!-- AdminLTE App -->
+    <script src="/backend/adminlte/dist/js/app.min.js"></script>
 
-        // Set active state on menu element
-        var current_url = "{{ url(Route::current()->getUri()) }}";
-        $("ul.sidebar-menu li a").each(function() {
-          if ($(this).attr('href').startsWith(current_url) || current_url.startsWith($(this).attr('href')))
-          {
-            $(this).parents('li').addClass('active');
-          }
-        });
-    </script>
+    <!-- Optionally, you can add Slimscroll and FastClick plugins.
+         Both of these plugins are recommended to enhance the
+         user experience. Slimscroll is required when using the
+         fixed layout. -->
 
     @include('bytenet::inc.alerts')